restructure code and sunset non-multiresolution support
