+++ requirements/requirements-core.txt
@@ -4,5 +4,5 @@ colorlog
 pyyaml==6.0
 scipy
 
-# Switch to pypi install when available
-s2fft @ git+https://github.com/astro-informatics/s2fft.git@main#egg=s2fft
+# For spherical transforms
+s2fft >= 1.0.2
