+++ s2wav/__init__.py
@@ -1,1 +1,3 @@
-from .transforms.jax_wavelets import analysis, synthesis, flm_to_analysis
+from . import filters
+from . import samples
+from .transforms.rec_wav_jax import analysis, synthesis, flm_to_analysis
