+++ s2wav/transforms/rec_wav_jax.py
@@ -1,78 +1,12 @@
 from jax import jit
 import jax.numpy as jnp
-from s2wav.utils import shapes
 from functools import partial
 from typing import Tuple, List
-from s2fft.precompute_transforms.construct import (
-    wigner_kernel_jax,
-    spin_spherical_kernel_jax,
-)
-from s2fft.precompute_transforms import wigner, spherical
+import s2fft
+from s2wav import samples
+from s2wav.transforms import construct
 
-
-def generate_precomputes(
-    L: int,
-    N: int,
-    J_min: int = 0,
-    lam: float = 2.0,
-    sampling: str = "mw",
-    nside: int = None,
-    forward: bool = False,
-    reality: bool = False,
-    multiresolution: bool = False,
-    nospherical: bool = False,
-) -> List[jnp.ndarray]:
-    r"""Generates a list of precompute arrays associated with the underlying Wigner
-    transforms.
-
-    Args:
-        L (int): Harmonic bandlimit.
-
-        N (int, optional): Upper azimuthal band-limit. Defaults to 1.
-
-        J_min (int, optional): Lowest frequency wavelet scale to be used. Defaults to 1.
-
-        lam (float, optional): Wavelet parameter which determines the scale factor between consecutive wavelet scales.
-            Note that :math:`\lambda = 2` indicates dyadic wavelets. Defaults to 2.
-
-        sampling (str, optional): Spherical sampling scheme from {"mw","mwss", "dh",
-            "healpix"}. Defaults to "mw".
-
-        nside (int, optional): HEALPix Nside resolution parameter.  Only required if sampling="healpix".  Defaults
-            to None.
-
-        forward (bool, optional): _description_. Defaults to False.
-
-        reality (bool, optional): Whether :math:`f \in \mathbb{R}`, if True exploits
-            conjugate symmetry of harmonic coefficients. Defaults to False.
-
-        multiresolution (bool, optional): Whether to store the scales at :math:`j_{\text{max}}`
-            resolution or its own resolution. Defaults to False.
-
-        nospherical (bool, optional): Whether to only compute Wigner precomputes.
-            Defaults to False.
-
-    Returns:
-        List[jnp.ndarray]: Precomputed recursion arrays for underlying Wigner transforms.
-    """
-    precomps = []
-    J = shapes.j_max(L, lam)
-    for j in range(J_min, J + 1):
-        Lj, Nj, L0j = shapes.LN_j(L, j, N, lam, multiresolution)
-        precomps.append(wigner_kernel_jax(Lj, Nj, reality, sampling, nside, forward))
-    Ls = shapes.scal_bandlimit(L, J_min, lam, multiresolution)
-    if nospherical:
-        return [], [], precomps
-    precompute_scaling = spin_spherical_kernel_jax(
-        Ls, 0, reality, sampling, nside, forward
-    )
-    precompute_full = spin_spherical_kernel_jax(
-        L, 0, reality, sampling, nside, not forward
-    )
-    return precompute_full, precompute_scaling, precomps
-
-
-@partial(jit, static_argnums=(2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 13))
+@partial(jit, static_argnums=(2, 3, 4, 5, 6, 7, 8, 9))
 def synthesis(
     f_wav: jnp.ndarray,
     f_scal: jnp.ndarray,
@@ -81,13 +15,10 @@ def synthesis(
     J_min: int = 0,
     lam: float = 2.0,
     spin: int = 0,
-    spin0: int = 0,
     sampling: str = "mw",
     nside: int = None,
     reality: bool = False,
-    multiresolution: bool = False,
     filters: Tuple[jnp.ndarray] = None,
-    spmd: bool = False,
     precomps: List[List[jnp.ndarray]] = None,
 ) -> jnp.ndarray:
     r"""Computes the synthesis directional wavelet transform [1,2].
@@ -110,8 +41,6 @@ def synthesis(
 
         spin (int, optional): Spin (integer) of input signal. Defaults to 0.
 
-        spin0 (int, optional): Spin (integer) of output signal. Defaults to 0.
-
         sampling (str, optional): Spherical sampling scheme from {"mw","mwss", "dh",
             "healpix"}. Defaults to "mw".
 
@@ -121,14 +50,7 @@ def synthesis(
         reality (bool, optional): Whether :math:`f \in \mathbb{R}`, if True exploits
             conjugate symmetry of harmonic coefficients. Defaults to False.
 
-        multiresolution (bool, optional): Whether to store the scales at :math:`j_{\text{max}}`
-            resolution or its own resolution. Defaults to False.
-
         filters (Tuple[jnp.ndarray], optional): Precomputed wavelet filters. Defaults to None.
-
-        spmd (bool, optional): Whether to map compute over multiple devices. Currently this
-            only maps over all available devices, and is only valid for JAX implementations.
-            Defaults to False.
 
         precomps (List[jnp.ndarray]): Precomputed list of recursion coefficients. At most
             of length :math:`L^2`, which is a minimal memory overhead.
@@ -144,22 +66,27 @@ def synthesis(
         [2] J. McEwen et. al., "Directional spin wavelets on the sphere", arXiv preprint arXiv:1509.06749 (2015).
     """
     if precomps == None:
-        raise ValueError("Must provide precomputed kernels for this transform!")
-
-    J = shapes.j_max(L, lam)
-    Ls = shapes.scal_bandlimit(L, J_min, lam, multiresolution)
+        precomps = construct.generate_wigner_precomputes(
+            L, N, J_min, lam, sampling, nside, True, reality
+        )
+    J = samples.j_max(L, lam)
+    Ls = samples.scal_bandlimit(L, J_min, lam, True)
     flm = jnp.zeros((L, 2 * L - 1), dtype=jnp.complex128)
-    f_scal_lm = spherical.forward_transform_jax(
-        f_scal, precomps[1], Ls, sampling, reality, spin, nside
-    )
+    f_scal_lm = s2fft.forward_jax(f_scal, Ls, spin, nside, sampling, reality)
 
     # Sum the all wavelet wigner coefficients for each lmn
     # Note that almost the entire compute is concentrated at the highest J
     for j in range(J_min, J + 1):
-        Lj, Nj, L0j = shapes.LN_j(L, j, N, lam, multiresolution)
-        spmd_iter = spmd if N == Nj else False
-        temp = wigner.forward_transform_jax(
-            f_wav[j - J_min], precomps[2][j - J_min], Lj, Nj, sampling, reality, nside
+        Lj, Nj, L0j = samples.LN_j(L, j, N, lam, True)
+        temp = s2fft.wigner.forward_jax(
+            f_wav[j - J_min],
+            Lj,
+            Nj,
+            nside,
+            sampling,
+            reality,
+            precomps[j - J_min],
+            L_lower=L0j,
         )
         flm = flm.at[L0j:Lj, L - Lj : L - 1 + Lj].add(
             jnp.einsum(
@@ -175,12 +102,11 @@ def synthesis(
     flm = flm.at[:Ls, L - Ls : L - 1 + Ls].add(
         jnp.einsum("lm,l->lm", f_scal_lm, phi, optimize=True)
     )
-    return spherical.inverse_transform_jax(
-        flm, precomps[0], L, sampling, reality, spin, nside
-    )
+
+    return s2fft.inverse_jax(flm, L, spin, nside, sampling, reality)
 
 
-@partial(jit, static_argnums=(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 12))
+@partial(jit, static_argnums=(1, 2, 3, 4, 5, 6, 7, 8))
 def analysis(
     f: jnp.ndarray,
     L: int,
@@ -188,13 +114,10 @@ def analysis(
     J_min: int = 0,
     lam: float = 2.0,
     spin: int = 0,
-    spin0: int = 0,
     sampling: str = "mw",
     nside: int = None,
     reality: bool = False,
-    multiresolution: bool = False,
     filters: Tuple[jnp.ndarray] = None,
-    spmd: bool = False,
     precomps: List[List[jnp.ndarray]] = None,
 ) -> Tuple[jnp.ndarray]:
     r"""Wavelet analysis from pixel space to wavelet space for complex signals.
@@ -213,8 +136,6 @@ def analysis(
 
         spin (int, optional): Spin (integer) of input signal. Defaults to 0.
 
-        spin0 (int, optional): Spin (integer) of output signal. Defaults to 0.
-
         sampling (str, optional): Spherical sampling scheme from {"mw","mwss", "dh", "healpix"}. Defaults to "mw".
 
         nside (int, optional): HEALPix Nside resolution parameter.  Only required if sampling="healpix".  Defaults
@@ -223,14 +144,7 @@ def analysis(
         reality (bool, optional): Whether :math:`f \in \mathbb{R}`, if True exploits
             conjugate symmetry of harmonic coefficients. Defaults to False.
 
-        multiresolution (bool, optional): Whether to store the scales at :math:`j_{\text{max}}`
-            resolution or its own resolution. Defaults to False.
-
         filters (Tuple[jnp.ndarray], optional): Precomputed wavelet filters. Defaults to None.
-
-        spmd (bool, optional): Whether to map compute over multiple devices. Currently this
-            only maps over all available devices, and is only valid for JAX implementations.
-            Defaults to False.
 
         precomps (List[jnp.ndarray]): Precomputed list of recursion coefficients. At most
             of length :math:`L^2`, which is a minimal memory overhead.
@@ -243,13 +157,14 @@ def analysis(
             with shape :math:`[n_{\theta}, n_{\phi}]`.
     """
     if precomps == None:
-        raise ValueError("Must provide precomputed kernels for this transform!")
+        precomps = construct.generate_wigner_precomputes(
+            L, N, J_min, lam, sampling, nside, False, reality
+        )
+    J = samples.j_max(L, lam)
+    Ls = samples.scal_bandlimit(L, J_min, lam, True)
 
-    J = shapes.j_max(L, lam)
-    Ls = shapes.scal_bandlimit(L, J_min, lam, multiresolution)
-
-    f_wav_lmn = shapes.construct_flmn_jax(L, N, J_min, lam, multiresolution)
-    f_wav = shapes.construct_f_jax(L, N, J_min, lam, sampling, nside, multiresolution)
+    f_wav_lmn = samples.construct_flmn_jax(L, N, J_min, lam, True)
+    f_wav = samples.construct_f_jax(L, N, J_min, lam, sampling, nside, True)
 
     wav_lm = jnp.einsum(
         "jln, l->jln",
@@ -258,14 +173,12 @@ def analysis(
         optimize=True,
     )
 
-    flm = spherical.forward_transform_jax(
-        f, precomps[0], L, sampling, reality, spin, nside
-    )
+    flm = s2fft.forward_jax(f, L, spin, nside, sampling, reality)
+
     # Project all wigner coefficients for each lmn onto wavelet coefficients
     # Note that almost the entire compute is concentrated at the highest J
     for j in range(J_min, J + 1):
-        Lj, Nj, L0j = shapes.LN_j(L, j, N, lam, multiresolution)
-        spmd_iter = spmd if N == Nj else False
+        Lj, Nj, L0j = samples.LN_j(L, j, N, lam, True)
         f_wav_lmn[j - J_min] = (
             f_wav_lmn[j - J_min]
             .at[::2, L0j:]
@@ -279,14 +192,16 @@ def analysis(
             )
         )
 
-        f_wav[j - J_min] = wigner.inverse_transform_jax(
+        f_wav[j - J_min] = s2fft.wigner.inverse_jax(
             f_wav_lmn[j - J_min],
-            precomps[2][j - J_min],
             Lj,
             Nj,
+            nside,
             sampling,
             reality,
-            nside,
+            precomps[j - J_min],
+            False,
+            L0j,
         )
 
     # Project all harmonic coefficients for each lm onto scaling coefficients
@@ -296,13 +211,11 @@ def analysis(
     if Ls == 1:
         f_scal = temp * jnp.sqrt(1 / (4 * jnp.pi))
     else:
-        f_scal = spherical.inverse_transform_jax(
-            temp, precomps[1], Ls, sampling, reality, spin, nside
-        )
+        f_scal = s2fft.inverse_jax(temp, Ls, spin, nside, sampling, reality)
     return f_wav, f_scal
 
 
-@partial(jit, static_argnums=(1, 2, 3, 4, 5, 6, 7, 8, 9, 11))
+@partial(jit, static_argnums=(1, 2, 3, 4, 5, 6, 7, 8))
 def flm_to_analysis(
     flm: jnp.ndarray,
     L: int,
@@ -313,9 +226,7 @@ def flm_to_analysis(
     sampling: str = "mw",
     nside: int = None,
     reality: bool = False,
-    multiresolution: bool = False,
     filters: Tuple[jnp.ndarray] = None,
-    spmd: bool = False,
     precomps: List[List[jnp.ndarray]] = None,
 ) -> Tuple[jnp.ndarray]:
     r"""Wavelet analysis from pixel space to wavelet space for complex signals.
@@ -332,10 +243,6 @@ def flm_to_analysis(
         lam (float, optional): Wavelet parameter which determines the scale factor between consecutive wavelet scales.
             Note that :math:`\lambda = 2` indicates dyadic wavelets. Defaults to 2.
 
-        spin (int, optional): Spin (integer) of input signal. Defaults to 0.
-
-        spin0 (int, optional): Spin (integer) of output signal. Defaults to 0.
-
         sampling (str, optional): Spherical sampling scheme from {"mw","mwss", "dh", "healpix"}. Defaults to "mw".
 
         nside (int, optional): HEALPix Nside resolution parameter.  Only required if sampling="healpix".  Defaults
@@ -344,14 +251,7 @@ def flm_to_analysis(
         reality (bool, optional): Whether :math:`f \in \mathbb{R}`, if True exploits
             conjugate symmetry of harmonic coefficients. Defaults to False.
 
-        multiresolution (bool, optional): Whether to store the scales at :math:`j_{\text{max}}`
-            resolution or its own resolution. Defaults to False.
-
-        filters (Tuple[jnp.ndarray], optional): Precomputed wavelet filters. Defaults to None.
-
-        spmd (bool, optional): Whether to map compute over multiple devices. Currently this
-            only maps over all available devices, and is only valid for JAX implementations.
-            Defaults to False.
+        filters (jnp.ndarray, optional): Precomputed wavelet filters. Defaults to None.
 
         precomps (List[jnp.ndarray]): Precomputed list of recursion coefficients. At most
             of length :math:`L^2`, which is a minimal memory overhead.
@@ -359,17 +259,16 @@ def flm_to_analysis(
     Returns:
         f_wav (jnp.ndarray): Array of wavelet pixel-space coefficients
             with shape :math:`[n_{J}, 2N-1, n_{\theta}, n_{\phi}]`.
-
-        f_scal (jnp.ndarray): Array of scaling pixel-space coefficients
-            with shape :math:`[n_{\theta}, n_{\phi}]`.
     """
     if precomps == None:
-        raise ValueError("Must provide precomputed kernels for this transform!")
+        precomps = construct.generate_wigner_precomputes(
+            L, N, J_min, lam, sampling, nside, False, reality
+        )
 
-    J = J_max if J_max is not None else shapes.j_max(L, lam)
+    J = J_max if J_max is not None else samples.j_max(L, lam)
 
-    f_wav_lmn = shapes.construct_flmn_jax(L, N, J_min, lam, multiresolution)
-    f_wav = shapes.construct_f_jax(L, N, J_min, lam, sampling, nside, multiresolution)
+    f_wav_lmn = samples.construct_flmn_jax(L, N, J_min, lam, True)
+    f_wav = samples.construct_f_jax(L, N, J_min, lam, sampling, nside, True)
 
     wav_lm = jnp.einsum(
         "jln, l->jln",
@@ -381,7 +280,7 @@ def flm_to_analysis(
     # Project all wigner coefficients for each lmn onto wavelet coefficients
     # Note that almost the entire compute is concentrated at the highest J
     for j in range(J_min, J + 1):
-        Lj, Nj, L0j = shapes.LN_j(L, j, N, lam, multiresolution)
+        Lj, Nj, L0j = samples.LN_j(L, j, N, lam, True)
         f_wav_lmn[j - J_min] = (
             f_wav_lmn[j - J_min]
             .at[::2, L0j:]
@@ -395,14 +294,16 @@ def flm_to_analysis(
             )
         )
 
-        f_wav[j - J_min] = wigner.inverse_transform_jax(
+        f_wav[j - J_min] = s2fft.wigner.inverse_jax(
             f_wav_lmn[j - J_min],
-            precomps[2][j - J_min],
             Lj,
             Nj,
+            nside,
             sampling,
             reality,
-            nside,
+            precomps[j - J_min],
+            False,
+            L0j,
         )
 
     return f_wav
