--- s2wav/utils/__init__.py
@@ -1,2 +1,0 @@
-from . import math_functions
-from . import shapes
