+++ tests/test_gradients.py
@@ -1,12 +1,9 @@
 import pytest
-import numpy as np
-
-from s2wav.transforms import jax_wavelets
-from s2wav.filter_factory import filters
-from s2wav.utils import shapes
 import jax.numpy as jnp
 from jax.test_util import check_grads
 import s2fft
+from s2wav.transforms import rec_wav_jax, pre_wav_jax, construct
+from s2wav import filters, samples
 
 L_to_test = [8]
 N_to_test = [3]
@@ -14,13 +11,14 @@ J_min_to_test = [2]
 multiresolution = [False, True]
 reality = [False, True]
 sampling_to_test = ["mw", "mwss", "dh"]
-
+recursive_transform = [False, True]
 
 @pytest.mark.parametrize("L", L_to_test)
 @pytest.mark.parametrize("N", N_to_test)
 @pytest.mark.parametrize("J_min", J_min_to_test)
 @pytest.mark.parametrize("multiresolution", multiresolution)
 @pytest.mark.parametrize("reality", reality)
+@pytest.mark.parametrize("recursive", recursive_transform)
 def test_jax_synthesis_gradients(
     flm_generator,
     L: int,
@@ -28,18 +26,29 @@ def test_jax_synthesis_gradients(
     J_min: int,
     multiresolution: bool,
     reality: bool,
+    recursive: bool
 ):
-    J = shapes.j_max(L)
+    J = samples.j_max(L)
     if J_min >= J:
         pytest.skip("J_min larger than J which isn't a valid test case.")
 
     # Generate wavelet filters
     filter = filters.filters_directional_vectorised(L, N, J_min)
+    generator = construct.generate_wigner_precomputes if recursive else construct.generate_full_precomputes
+    synthesis = rec_wav_jax.synthesis if recursive else pre_wav_jax.synthesis
+    precomps = generator(
+        L,
+        N,
+        J_min,
+        forward=True,
+        reality=reality,
+        multiresolution=multiresolution,
+    )
 
     # Generate random signal
     flm = flm_generator(L=L, L_lower=0, spin=0, reality=reality)
     f = s2fft.inverse_jax(flm, L)
-    f_wav, f_scal = jax_wavelets.analysis(
+    f_wav, f_scal = rec_wav_jax.analysis(
         f,
         L,
         N,
@@ -54,7 +63,7 @@ def test_jax_synthesis_gradients(
     f_target = s2fft.inverse_jax(flm_target, L)
 
     def func(f_wav, f_scal):
-        f = jax_wavelets.synthesis(
+        f = synthesis(
             f_wav,
             f_scal,
             L,
@@ -63,6 +72,7 @@ def test_jax_synthesis_gradients(
             multiresolution=multiresolution,
             reality=reality,
             filters=filter,
+            precomps=precomps,
         )
         return jnp.sum(jnp.abs(f - f_target) ** 2)
 
@@ -82,6 +92,7 @@ def test_jax_synthesis_gradients(
 @pytest.mark.parametrize("J_min", J_min_to_test)
 @pytest.mark.parametrize("multiresolution", multiresolution)
 @pytest.mark.parametrize("reality", reality)
+@pytest.mark.parametrize("recursive", recursive_transform)
 def test_jax_analysis_gradients(
     flm_generator,
     L: int,
@@ -89,13 +100,24 @@ def test_jax_analysis_gradients(
     J_min: int,
     multiresolution: bool,
     reality: bool,
+    recursive: bool
 ):
-    J = shapes.j_max(L)
+    J = samples.j_max(L)
     if J_min >= J:
         pytest.skip("J_min larger than J which isn't a valid test case.")
 
     # Generate wavelet filters
     filter = filters.filters_directional_vectorised(L, N, J_min)
+    generator = construct.generate_wigner_precomputes if recursive else construct.generate_full_precomputes
+    analysis = rec_wav_jax.analysis if recursive else pre_wav_jax.analysis
+    precomps = generator(
+        L,
+        N,
+        J_min,
+        forward=False,
+        reality=reality,
+        multiresolution=multiresolution,
+    )
 
     # Generate random signal
     flm = flm_generator(L=L, L_lower=0, spin=0, reality=reality)
@@ -104,7 +126,7 @@ def test_jax_analysis_gradients(
     # Generate target signal
     flm_target = flm_generator(L=L, L_lower=0, spin=0, reality=reality)
     f_target = s2fft.inverse_jax(flm_target, L)
-    f_wav_target, f_scal_target = jax_wavelets.analysis(
+    f_wav_target, f_scal_target = rec_wav_jax.analysis(
         f_target,
         L,
         N,
@@ -115,7 +137,7 @@ def test_jax_analysis_gradients(
     )
 
     def func(f):
-        f_wav, f_scal = jax_wavelets.analysis(
+        f_wav, f_scal = analysis(
             f,
             L,
             N,
@@ -123,6 +145,7 @@ def test_jax_analysis_gradients(
             multiresolution=multiresolution,
             reality=reality,
             filters=filter,
+            precomps=precomps
         )
         loss = jnp.sum(jnp.abs(f_scal - f_scal_target) ** 2)
         for j in range(J - J_min):
