+++ tests/test_wavelets.py
@@ -1,40 +1,34 @@
 import pytest
 import numpy as np
 import pys2let as s2let
-
-from s2wav.transforms import jax_wavelets
-from s2wav.filter_factory import filters
-from s2wav.utils import shapes
-from s2fft import base_transforms as base
+from s2fft import base_transforms as sht_base
+from s2wav.transforms import rec_wav_jax, pre_wav_jax, construct
+from s2wav import filters, samples
 
 L_to_test = [8]
 N_to_test = [2, 3]
 J_min_to_test = [2]
 lam_to_test = [2, 3]
-multiresolution = [False, True]
 reality = [False, True]
-multiple_gpus = [False]
 sampling_to_test = ["mw", "mwss", "dh"]
-
+recursive_transform = [False, True]
 
 @pytest.mark.parametrize("L", L_to_test)
 @pytest.mark.parametrize("N", N_to_test)
 @pytest.mark.parametrize("J_min", J_min_to_test)
 @pytest.mark.parametrize("lam", lam_to_test)
-@pytest.mark.parametrize("multiresolution", multiresolution)
 @pytest.mark.parametrize("reality", reality)
-@pytest.mark.parametrize("spmd", multiple_gpus)
+@pytest.mark.parametrize("recursive", recursive_transform)
 def test_jax_synthesis(
     wavelet_generator,
     L: int,
     N: int,
     J_min: int,
     lam: int,
-    multiresolution: bool,
     reality: bool,
-    spmd: bool,
+    recursive: bool
 ):
-    J = shapes.j_max(L, lam)
+    J = samples.j_max(L, lam)
     if J_min >= J:
         pytest.skip("J_min larger than J which isn't a valid test case.")
 
@@ -43,7 +37,6 @@ def test_jax_synthesis(
         N=N,
         J_min=J_min,
         lam=lam,
-        multiresolution=multiresolution,
         reality=reality,
     )
 
@@ -55,32 +48,31 @@ def test_jax_synthesis(
         J_min,
         N,
         spin=0,
-        upsample=not multiresolution,
+        upsample=False,
     )
 
-    # Precompute some values
     filter = filters.filters_directional_vectorised(L, N, J_min, lam)
-    precomps = jax_wavelets.generate_wigner_precomputes(
+    generator = construct.generate_wigner_precomputes if recursive else construct.generate_full_precomputes
+    synthesis = rec_wav_jax.synthesis if recursive else pre_wav_jax.synthesis
+
+    precomps = generator(
         L,
         N,
         J_min,
         lam,
         forward=True,
-        reality=reality,
-        multiresolution=multiresolution,
-    )
-    f_check = jax_wavelets.synthesis(
+        reality=reality
+    ) 
+    f_check = synthesis(
         f_wav,
         f_scal,
         L,
         N,
         J_min,
         lam,
-        multiresolution=multiresolution,
         reality=reality,
         filters=filter,
-        precomps=precomps,
-        spmd=spmd,
+        precomps=precomps
     )
     f = np.real(f) if reality else f
     np.testing.assert_allclose(f, f_check.flatten("C"), atol=1e-14)
@@ -90,9 +82,8 @@ def test_jax_synthesis(
 @pytest.mark.parametrize("N", N_to_test)
 @pytest.mark.parametrize("J_min", J_min_to_test)
 @pytest.mark.parametrize("lam", lam_to_test)
-@pytest.mark.parametrize("multiresolution", multiresolution)
 @pytest.mark.parametrize("reality", reality)
-@pytest.mark.parametrize("spmd", multiple_gpus)
+@pytest.mark.parametrize("recursive", recursive_transform)
 def test_jax_analysis(
     flm_generator,
     f_wav_converter,
@@ -100,16 +91,15 @@ def test_jax_analysis(
     N: int,
     J_min: int,
     lam: int,
-    multiresolution: bool,
     reality: bool,
-    spmd: bool,
+    recursive: bool
 ):
-    J = shapes.j_max(L, lam)
+    J = samples.j_max(L, lam)
     if J_min >= J:
         pytest.skip("J_min larger than J which isn't a valid test case.")
 
     flm = flm_generator(L=L, L_lower=0, spin=0, reality=reality)
-    f = base.spherical.inverse(flm, L, reality=reality)
+    f = sht_base.spherical.inverse(flm, L, reality=reality)
 
     f_wav, f_scal = s2let.analysis_px2wav(
         f.flatten("C").astype(np.complex128),
@@ -118,32 +108,31 @@ def test_jax_analysis(
         J_min,
         N,
         spin=0,
-        upsample=not multiresolution,
+        upsample=False
     )
     filter = filters.filters_directional_vectorised(L, N, J_min, lam)
-    precomps = jax_wavelets.generate_wigner_precomputes(
+    generator = construct.generate_wigner_precomputes if recursive else construct.generate_full_precomputes
+    analysis = rec_wav_jax.analysis if recursive else pre_wav_jax.analysis
+    precomps = generator(
         L,
         N,
         J_min,
         lam,
         forward=False,
-        reality=reality,
-        multiresolution=multiresolution,
+        reality=reality
     )
-    f_wav_check, f_scal_check = jax_wavelets.analysis(
+    f_wav_check, f_scal_check = analysis(
         f,
         L,
         N,
         J_min,
         lam,
-        multiresolution=multiresolution,
         reality=reality,
         filters=filter,
-        precomps=precomps,
-        spmd=spmd,
+        precomps=precomps
     )
 
-    f_wav_check = f_wav_converter(f_wav_check, L, N, J_min, lam, multiresolution)
+    f_wav_check = f_wav_converter(f_wav_check, L, N, J_min, lam, True)
 
     np.testing.assert_allclose(f_wav, f_wav_check, atol=1e-14)
     np.testing.assert_allclose(f_scal, f_scal_check.flatten("C"), atol=1e-14)
@@ -153,53 +142,45 @@ def test_jax_analysis(
 @pytest.mark.parametrize("N", N_to_test)
 @pytest.mark.parametrize("J_min", J_min_to_test)
 @pytest.mark.parametrize("lam", lam_to_test)
-@pytest.mark.parametrize("multiresolution", multiresolution)
 @pytest.mark.parametrize("reality", reality)
 @pytest.mark.parametrize("sampling", sampling_to_test)
-@pytest.mark.parametrize("spmd", multiple_gpus)
 def test_jax_round_trip(
     flm_generator,
     L: int,
     N: int,
     J_min: int,
     lam: int,
-    multiresolution: bool,
     reality: bool,
-    sampling: str,
-    spmd: bool,
+    sampling: str
 ):
-    J = shapes.j_max(L, lam)
+    J = samples.j_max(L, lam)
     if J_min >= J:
         pytest.skip("J_min larger than J which isn't a valid test case.")
 
     flm = flm_generator(L=L, L_lower=0, spin=0, reality=reality)
-    f = base.spherical.inverse(flm, L, reality=reality, sampling=sampling)
+    f = sht_base.spherical.inverse(flm, L, reality=reality, sampling=sampling)
     filter = filters.filters_directional_vectorised(L, N, J_min, lam)
 
-    f_wav, f_scal = jax_wavelets.analysis(
+    f_wav, f_scal = rec_wav_jax.analysis(
         f,
         L,
         N,
         J_min,
         lam,
-        multiresolution=multiresolution,
         reality=reality,
         sampling=sampling,
-        filters=filter,
-        spmd=spmd,
+        filters=filter
     )
-    f_check = jax_wavelets.synthesis(
+    f_check = rec_wav_jax.synthesis(
         f_wav,
         f_scal,
         L,
         N,
         J_min,
         lam,
-        multiresolution=multiresolution,
         sampling=sampling,
         reality=reality,
-        filters=filter,
-        spmd=spmd,
+        filters=filter
     )
 
     np.testing.assert_allclose(f, f_check, atol=1e-14)
